requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.0
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
PyMuPDF==1.26.3
pytesseract==0.3.13
python-dotenv==1.1.1
python-pptx==1.0.2
typing_extensions==4.14.1
uuid==1.30
Werkzeug==3.1.3
xlsxwriter==3.2.5
